requirements.txt
viam-sdk
transformers
pillow
torch
torchvision